extract shared behavior to a trait
